refactor(web): move FileData type to types.ts to fix type-aware lint
